Fixed missing constructor bug
Hibernate and @Builder need specific constructors.
If I only use @Data, it will create a NoArgsCons. @Builder will not create a AllArgsCons if there is another constructor annotation.
Hibernate on the other hand needs a NoArgsCons, which will not be created by @Data if I explicitly create a AllArgsCons.
Solution: Explicitly create NonArgs and AllArgs constructors.